# src/main/resources/openapi/university-scheduler.yaml
openapi: 3.0.3
info:
  title: University Scheduler
  description: University Scheduler API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /teachers/extended:
    get:
      summary: Get all teachers with extended information
      operationId: findAllExtended
      tags:
        - Teacherz
      parameters:
        - $ref: '#/components/parameters/pageParams'
        - $ref: '#/components/parameters/universityCode'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeacherExtendedInfo'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /teachers/{id}:
    get:
      summary: Get teacher by id
      operationId: findById
      tags:
        - Teacherz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/universityCode'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherExtendedInfo'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update teacher
      operationId: update
      tags:
        - Teacherz
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherExtendedInfo'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherExtendedInfo'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete teacher by id
      operationId: deleteById
      tags:
        - Teacherz
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /subjects/extended:
    get:
      summary: Get all subjects with extended information
      operationId: findAllExtended
      tags:
        - Subjectz
      parameters:
        - $ref: '#/components/parameters/pageParams'
        - $ref: '#/components/parameters/universityCode'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectExtendedInfo'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /subjects/{id}:
    get:
      summary: Get subject by id
      operationId: findById
      tags:
        - Subjectz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/universityCode'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectExtendedInfo'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update subject
      operationId: update
      tags:
        - Subjectz
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectExtendedInfo'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectExtendedInfo'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete subject by id
      operationId: deleteById
      tags:
        - Subjectz
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /universities/{id}:
    get:
      summary: Get university by id
      operationId: findById
      tags:
        - Universitiez
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UniversityInfo'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update university
      operationId: update
      tags:
        - Universitiez
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UniversityInfo'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UniversityInfo'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete university by id
      operationId: deleteById
      tags:
        - Universitiez
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /educational-programs/{id}:
    get:
      summary: Get educational program by id
      operationId: findById
      tags:
        - EducationalProgramz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/universityCode'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EducationalProgramInfo'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update educational program
      operationId: update
      tags:
        - EducationalProgramz
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EducationalProgramInfo'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EducationalProgramInfo'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete educational program by id
      operationId: deleteById
      tags:
        - EducationalProgramz
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /groups/{id}:
    get:
      summary: Get group by id
      operationId: findById
      tags:
        - Groupz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/universityCode'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupInfo'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update group
      operationId: update
      tags:
        - Groupz
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupInfo'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupInfo'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete group
      operationId: deleteById
      tags:
        - Groupz
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /schedules/extended:
    get:
      summary: Get all extended schedules
      operationId: findAllExtended
      tags:
        - Schedulez
      parameters:
        - $ref: '#/components/parameters/searchId'
        - $ref: '#/components/parameters/searchType'
        - $ref: '#/components/parameters/pageParams'
        - $ref: '#/components/parameters/universityCode'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleExtendedInfo'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedules/{id}:
    get:
      summary: Get schedule by id
      operationId: findById
      tags:
        - Schedulez
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/universityCode'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleInfo'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update schedule
      operationId: update
      tags:
        - Schedulez
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleInfo'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleInfo'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete schedule
      operationId: deleteById
      tags:
        - Schedulez
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    pageParams:
      in: query
      name: pageParams
      description: Page parameters
      required: false
      schema:
        $ref: '#/components/schemas/PageParams'
...
    searchId:
      in: query
      name: searchId
      description: Search id
      required: false
      schema:
        $ref: '#/components/schemas/SearchId'
    searchType:
      in: query
      name: searchType
      description: Searchtype
      required: false
      schema:
        $ref: '#/components/schemas/SearchType'
    universityCode:
      in: query
      name: universityCode
      description: University code
      required: false
      schema:
        $ref: '#/components/schemas/UniversityCode'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PageParams:
      description: Page parameters
      type: object
      properties:
        pageSize:
          description: Page size
          type: integer
          minimum: 1
          maximum: 50
        pageCurrent:
          description: Current page
          type: integer
          minimum: 1
        pageTotal:
          description: Total pages
          type: integer
    UniversityCode:
      description: University code
      type: string
...
    TeacherExtendedInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        qualification:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        department:
          type: string
...
    SubjectExtendedInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        code:
          type: string
        description:
          type: string
        credits:
          type: integer
          format: int32
    UniversityInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        userId:
          type: string
          format: uuid
    Error:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
    EducationalProgramInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        elective:
          type: boolean
        universityId:
          type: string
          format: uuid
...
    GroupInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        course:
          type: integer
          format: int32
        educationalProgramId:
          type: string
          format: uuid
    ScheduleInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sessionType:
          $ref: '#/components/schemas/SessionType'
        subjectId:
          type: string
          format: uuid
        teacherId:
          type: string
          format: uuid
        groups:
          type: array
          items:
            type: string
            format: uuid
        room:
          type: string
        week:
          $ref: '#/components/schemas/Week'
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    ScheduleExtendedInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sessionType:
          $ref: '#/components/schemas/SessionType'
        subjectName:
          type: string
        teacherName:
          type: string
        groups:
          type: array
          items:
            type: string
        room:
          type: string
        week:
          $ref: '#/components/schemas/Week'
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    Week:
      type: string
      description: Week day
      enum:
        - MONDAY
        - TUESDAY
        - WEDNESDAY
        - THURSDAY
        - FRIDAY
        - SATURDAY
        - SUNDAY
...
    SearchId:
      type: string
      format: uuid
    SearchType:
      type: string
      description: Search type
      enum:
        - TEACHER
        - SUBJECT
        - UNIVERSITY
        - EDUCATIONAL_PROGRAM
        - GROUP
...
    SessionType:
      type: string
      description: Session type
      enum:
        - LECTURE
        - LABORATORY_WORK
        - PRACTICE
...